Function bit_reversal, static_bit_reversal template and some utilities declared in integer/detail/utils.hpp done.
SVN-Revision: 62363

+++ SOC/2010/bits_and_ints/libs/integer/test/sign_extend_test.cpp
@@ -3,7 +3,7 @@
 //  (C) Copyright Murilo Adriano Vasconcelos 2010.
 //  Distributed under the Boost Software License, Version 1.0. (See
 //  accompanying file LICENSE_1_0.txt or copy at
-//  http://www.boost.org/LICENSE_1_0.txt)
+//  http://www.boost.org/LICENSE_1_0.txt
 
 
 #include <boost/detail/lightweight_test.hpp>
@@ -65,7 +65,7 @@ int main(int, char* [])
 #ifndef BOOST_NO_INT64_T
 	SIGN_EXTEND_TEST(0x7FFFFFFFFF, 39, 0xFFFFFFFFFFFFFFFF);
 	SIGN_EXTEND_TEST(0x1000000000000000, 63, 0x1000000000000000);
-	SIGN_EXTEND_TEST(0x7FFFFFFFFFFFFFFF, 63, -0x1);
+	SIGN_EXTEND_TEST(0x7FFFFFFFFFFFFFFF, 63, (uint64_t)-0x1);
 #endif
 	
 	return boost::report_errors();
